Start completely rebuilding the "safe output file" functionality
The current implementation produces a file object with extra methods/traces attached. This can be convenient, but is brittle and causes problems down the line because SafeOutputFile instances are not perfect proxies/emulators for the real streams produced by open(), causing problems in certain situations (e.g. not recognized as writable or originating in a file etc.)

While I could resort to using a more professional proxying package (wrapt etc) that could address some of the problems, this would add a heavy dependency to what is supposed to be a very basic, lightweight library. I'd rather alter the API to something that is a bit less fancy, but more robust.